Moved CreateCouponActivity.java
