+++ app/src/main/res/layout/activity_create_coupon.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".CreateCouponActivity">
+    tools:context=".coupon.CreateCouponActivity">
 
     <android.support.v7.widget.Toolbar
         android:id="@+id/toolbar"
