+++ app/src/main/res/menu/create_coupon.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".CreateCouponActivity">
+    tools:context=".coupon.CreateCouponActivity">
 
     <item
         android:id="@+id/menu_send"
